cid#1660814 Dereference null return value
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/188715
Tested-by: Jenkins
(cherry picked from commit 3e0e4d141bc0bb1f9ca5cd7766cd64c7ae16d4f5)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/189051
(cherry picked from commit d5e73fee5fecceaf711cb2f74eb1ff7a26f2be29)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/189076

+++ sd/source/ui/view/drviews4.cxx
@@ -166,8 +166,8 @@ void DrawViewShell::DeleteActualLayer()
                                                    aString));
     if (xQueryBox->run() == RET_YES)
     {
-        const SdrLayer* pLayer = rAdmin.GetLayer(aName);
-        mpDrawView->DeleteLayer( pLayer->GetName() );
+        if (const SdrLayer* pLayer = rAdmin.GetLayer(aName))
+            mpDrawView->DeleteLayer(pLayer->GetName());
 
         /* in order to redraw TabBar and Window; should be initiated later on by
            a hint from Joe (as by a change if the layer order). */
